Add context manager and exceptions handling

+++ data_collection/google_images/google_images_downloader.py
@@ -12,16 +12,22 @@ class GoogleImagesDownloader:
         self._images_per_artist = images_per_artist
         self._invalid_os_chars_regex = re.compile(r'[/:\"]')
         self._images_paths = None
+        self._clear_images_dir()
 
     def download(self, artist: str) -> List[str]:
-        self._crawl(artist)
+        try:
+            return self._download(artist)
+
+        except:
+            print('Received exception! returning empty list')
+            return []
+
+    def _download(self, artist: str) -> List[str]:
+        crawler = GoogleImageCrawler(storage={'root_dir': GOOGLE_IMAGES_RESOURCES_PATH})
+        crawler.crawl(keyword=f'{artist} musician', max_num=self._images_per_artist)
         self._images_paths = self._get_image_paths()
 
         return self._images_paths
-
-    def _crawl(self, artist: str) -> None:
-        crawler = GoogleImageCrawler(storage={'root_dir': GOOGLE_IMAGES_RESOURCES_PATH})
-        crawler.crawl(keyword=f'{artist} musician', max_num=self._images_per_artist)
 
     @staticmethod
     def _get_image_paths() -> List[str]:
@@ -35,10 +41,17 @@ class GoogleImagesDownloader:
 
         return images_paths
 
+    def _clear_images_dir(self) -> None:
+        for image_path in self._get_image_paths():
+            os.remove(image_path)
+
     def __enter__(self):
         return self
 
     def __exit__(self, exc_type, exc_val, exc_tb):
+        if self._images_paths is None:
+            return
+
         for image_path in self._images_paths:
             os.remove(image_path)
